requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.2
django-nested-admin==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
idna==3.3
inflection==0.3.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==6.2.5
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-drf==1.1.2
pytest-fixture-order==0.1.3
pytest-lambda==1.2.5
python-decouple==3.6
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.3
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
